+++ R/SelfControlledCohort.R
@@ -27,6 +27,7 @@
 
 
 computeIrrs <- function(estimates) {
+
   computeIrr <- function(numOutcomesExposed, numOutcomesUnexposed, timeAtRiskExposed, timeAtRiskUnexposed) {
     test <- rateratio.test::rateratio.test(x = c(numOutcomesExposed,
                                                  numOutcomesUnexposed),
@@ -34,15 +35,178 @@ computeIrrs <- function(estimates) {
                                                  timeAtRiskUnexposed))
     return(c(test$estimate[1], test$conf.int))
   }
+
   irrs <- mapply(computeIrr,
                  numOutcomesExposed = estimates$numOutcomesExposed,
                  numOutcomesUnexposed = estimates$numOutcomesUnexposed,
                  timeAtRiskExposed = estimates$timeAtRiskExposed,
                  timeAtRiskUnexposed = estimates$timeAtRiskUnexposed)
-  estimates$irr <- irrs[1, ]
-  estimates$irrLb95 <- irrs[2, ]
-  estimates$irrUb95 <- irrs[3, ]
+  estimates$irr <- irrs[1,]
+  estimates$irrLb95 <- irrs[2,]
+  estimates$irrUb95 <- irrs[3,]
   return(estimates)
+}
+
+#' @title
+#' Comnpute time at risk exposed and time at risk unexposed for risk window parameters
+#' @param connection                       DatabaseConnector connection instance
+#' @param cdmDatabaseSchema                Name of database schema that contains the OMOP CDM and
+#'                                         vocabulary.
+#' @param cdmVersion                       Define the OMOP CDM version used: currently support "4" and
+#'                                         "5".
+#' @param oracleTempSchema                 For Oracle only: the name of the database schema where you
+#'                                         want all temporary tables to be managed. Requires
+#'                                         create/insert permissions to this database.
+#' @param exposureIds                      A vector containing the drug_concept_ids or
+#'                                         cohort_definition_ids of the exposures of interest. If empty,
+#'                                         all exposures in the exposure table will be included.
+#' @param outcomeIds                       The condition_concept_ids or cohort_definition_ids of the
+#'                                         outcomes of interest. If empty, all the outcomes in the
+#'                                         outcome table will be included.
+#' @param exposureDatabaseSchema           The name of the database schema that is the location where
+#'                                         the exposure data used to define the exposure cohorts is
+#'                                         available. If exposureTable = DRUG_ERA,
+#'                                         exposureDatabaseSchema is not used by assumed to be
+#'                                         cdmSchema.  Requires read permissions to this database.
+#' @param exposureTable                    The tablename that contains the exposure cohorts.  If
+#'                                         exposureTable <> DRUG_ERA, then expectation is exposureTable
+#'                                         has format of COHORT table: cohort_concept_id, SUBJECT_ID,
+#'                                         COHORT_START_DATE, COHORT_END_DATE.
+#' @param outcomeDatabaseSchema            The name of the database schema that is the location where
+#'                                         the data used to define the outcome cohorts is available. If
+#'                                         exposureTable = CONDITION_ERA, exposureDatabaseSchema is not
+#'                                         used by assumed to be cdmSchema.  Requires read permissions
+#'                                         to this database.
+#' @param outcomeTable                     The tablename that contains the outcome cohorts.  If
+#'                                         outcomeTable <> CONDITION_OCCURRENCE, then expectation is
+#'                                         outcomeTable has format of COHORT table:
+#'                                         COHORT_DEFINITION_ID, SUBJECT_ID, COHORT_START_DATE,
+#'                                         COHORT_END_DATE.
+#' @param firstExposureOnly                If TRUE, only use first occurrence of each drug concept id
+#'                                         for each person
+#' @param minAge                           Integer for minimum allowable age.
+#' @param maxAge                           Integer for maximum allowable age.
+#' @param studyStartDate                   Date for minimum allowable data for index exposure. Date
+#'                                         format is 'yyyymmdd'.
+#' @param studyEndDate                     Date for maximum allowable data for index exposure. Date
+#'                                         format is 'yyyymmdd'.
+#' @param addLengthOfExposureExposed       If TRUE, use the duration from drugEraStart -> drugEraEnd as
+#'                                         part of timeAtRisk.
+#' @param riskWindowStartExposed           Integer of days to add to drugEraStart for start of
+#'                                         timeAtRisk (0 to include index date, 1 to start the day
+#'                                         after).
+#' @param riskWindowEndExposed             Additional window to add to end of exposure period (if
+#'                                         addLengthOfExposureExposed = TRUE, then add to exposure end
+#'                                         date, else add to exposure start date).
+#' @param addLengthOfExposureUnexposed     If TRUE, use the duration from exposure start -> exposure
+#'                                         end as part of timeAtRisk looking back before exposure
+#'                                         start.
+#' @param riskWindowEndUnexposed           Integer of days to add to exposure start for end of
+#'                                         timeAtRisk (0 to include index date, -1 to end the day
+#'                                         before).
+#' @param riskWindowStartUnexposed         Additional window to add to start of exposure period (if
+#'                                         addLengthOfExposureUnexposed = TRUE, then add to exposure
+#'                                         end date, else add to exposure start date).
+#' @param hasFullTimeAtRisk                If TRUE, restrict to people who have full time-at-risk
+#'                                         exposed and unexposed.
+#' @param riskWindowsTable                 String: optionally store the risk windows in a (non-temporary)
+#'                                         table. If it is not set the temporary table #risk_windows is
+#'                                         created. This may be a large table, consider deleting before
+#'                                         closing the conneciton.
+#' @param resultsDatabaseSchema            Schema to oputput results to. Ignored if riskWindowsTable is
+#'                                         temporary.
+#' @export
+computeSccRiskWindows <- function(connection,
+                                  cdmDatabaseSchema,
+                                  cdmVersion = 5,
+                                  oracleTempSchema = NULL,
+                                  exposureIds = '',
+                                  exposureDatabaseSchema = cdmDatabaseSchema,
+                                  exposureTable = "drug_era",
+                                  firstExposureOnly = TRUE,
+                                  minAge = "",
+                                  maxAge = "",
+                                  studyStartDate = "",
+                                  studyEndDate = "",
+                                  addLengthOfExposureExposed = TRUE,
+                                  riskWindowStartExposed = 1,
+                                  riskWindowEndExposed = 30,
+                                  addLengthOfExposureUnexposed = TRUE,
+                                  riskWindowEndUnexposed = -1,
+                                  riskWindowStartUnexposed = -30,
+                                  hasFullTimeAtRisk = FALSE,
+                                  washoutPeriod = 0,
+                                  followupPeriod = 0,
+                                  riskWindowsTable = "#risk_windows",
+                                  resultsDatabaseSchema = NULL) {
+  exposureTable <- tolower(exposureTable)
+
+  if (exposureTable == "drug_era") {
+    exposureStartDate <- "drug_era_start_date"
+    exposureEndDate <- "drug_era_end_date"
+    exposureId <- "drug_concept_id"
+    exposurePersonId <- "person_id"
+  } else if (exposureTable == "drug_exposure") {
+    exposureStartDate <- "drug_exposure_start_date"
+    exposureEndDate <- "drug_exposure_end_date"
+    exposureId <- "drug_concept_id"
+    exposurePersonId <- "person_id"
+  } else {
+    exposureStartDate <- "cohort_start_date"
+    exposureEndDate <- "cohort_end_date"
+    if (cdmVersion == "4") {
+      exposureId <- "cohort_concept_id"
+    } else {
+      exposureId <- "cohort_definition_id"
+    }
+    exposurePersonId <- "subject_id"
+  }
+
+  if (exposureIds != '') {
+    DatabaseConnector::insertTable(connection = connection,
+                                   tableName = "#scc_exposure_ids",
+                                   data = data.frame(exposure_id = exposureIds),
+                                   tempTable = TRUE)
+  }
+
+  if (riskWindowsTable != "#risk_windows") {
+    if (is.null(resultsDatabaseSchema))
+      stop("Risk windows table is not temporary and resultsDatabaseSchema is not set")
+    riskWindowsTable <- SqlRender::render("@results_database_schema.@risk_windows_table",
+                                          results_database_schema = resultsDatabaseSchema,
+                                          risk_windows_table = riskWindowsTable)
+  }
+
+  renderedSql <- SqlRender::loadRenderTranslateSql(sqlFilename = "ComputeSccRiskWindows.sql",
+                                                   packageName = "SelfControlledCohort",
+                                                   dbms = connection@dbms,
+                                                   oracleTempSchema = oracleTempSchema,
+                                                   cdm_database_schema = cdmDatabaseSchema,
+                                                   exposure_ids = exposureIds,
+                                                   exposure_database_schema = exposureDatabaseSchema,
+                                                   exposure_table = exposureTable,
+                                                   exposure_start_date = exposureStartDate,
+                                                   exposure_end_date = exposureEndDate,
+                                                   exposure_id = exposureId,
+                                                   exposure_person_id = exposurePersonId,
+                                                   first_exposure_only = firstExposureOnly,
+                                                   min_age = minAge,
+                                                   max_age = maxAge,
+                                                   study_start_date = studyStartDate,
+                                                   study_end_date = studyEndDate,
+                                                   add_length_of_exposure_exposed = addLengthOfExposureExposed,
+                                                   risk_window_start_exposed = riskWindowStartExposed,
+                                                   risk_window_end_exposed = riskWindowEndExposed,
+                                                   add_length_of_exposure_unexposed = addLengthOfExposureUnexposed,
+                                                   risk_window_end_unexposed = riskWindowEndUnexposed,
+                                                   risk_window_start_unexposed = riskWindowStartUnexposed,
+                                                   has_full_time_at_risk = hasFullTimeAtRisk,
+                                                   washout_window = washoutPeriod,
+                                                   followup_window = followupPeriod,
+                                                   risk_windows_table = riskWindowsTable)
+
+  ParallelLogger::logInfo("Computing time at risk exposed and unexposed windows")
+  DatabaseConnector::executeSql(connection, renderedSql)
 }
 
 #' @title
@@ -130,6 +294,14 @@ computeIrrs <- function(estimates) {
 #'                                         average absolute time between treatment and outcome. Note,
 #'                                         may add significant computation time on some database
 #'                                         engines.
+#' @param riskWindowsTable                 String: optionally store the risk windows in a (non-temporary)
+#'                                         table.
+#' @param resultsTable                     String: optionally store the summary results (number exposed/
+#'                                         unexposed patients per outcome-exposure pair) in a (non-temporary)
+#'                                         table. Note that this table does not store the rate ratios, only
+#'                                         the values required to calculate rate ratios.
+#' @param resultsDatabaseSchema                    Schema to oputput results to. Ignored if resultsTable and
+#'                                         riskWindowsTable are temporary.
 #' @param washoutPeriod                    Integer to define required time observed before exposure
 #'                                         start.
 #' @param followupPeriod                   Integer to define required time observed after exposure
@@ -154,8 +326,8 @@ runSelfControlledCohort <- function(connectionDetails,
                                     cdmDatabaseSchema,
                                     cdmVersion = 5,
                                     oracleTempSchema = NULL,
-                                    exposureIds,
-                                    outcomeIds,
+                                    exposureIds = '',
+                                    outcomeIds = '',
                                     exposureDatabaseSchema = cdmDatabaseSchema,
                                     exposureTable = "drug_era",
                                     outcomeDatabaseSchema = cdmDatabaseSchema,
@@ -176,35 +348,17 @@ runSelfControlledCohort <- function(connectionDetails,
                                     washoutPeriod = 0,
                                     followupPeriod = 0,
                                     computeTarDistribution = FALSE,
+                                    riskWindowsTable = "#risk_windows",
+                                    resultsTable = "#results",
+                                    resultsDatabaseSchema = NULL,
                                     computeThreads = 1) {
   if (riskWindowEndExposed < riskWindowStartExposed && !addLengthOfExposureExposed)
     stop("Risk window end (exposed) should be on or after risk window start")
   if (riskWindowEndUnexposed < riskWindowStartUnexposed && !addLengthOfExposureUnexposed)
     stop("Risk window end (unexposed) should be on or after risk window start")
   start <- Sys.time()
-  exposureTable <- tolower(exposureTable)
-  outcomeTable <- tolower(outcomeTable)
-  if (exposureTable == "drug_era") {
-    exposureStartDate <- "drug_era_start_date"
-    exposureEndDate <- "drug_era_end_date"
-    exposureId <- "drug_concept_id"
-    exposurePersonId <- "person_id"
-  } else if (exposureTable == "drug_exposure") {
-    exposureStartDate <- "drug_exposure_start_date"
-    exposureEndDate <- "drug_exposure_end_date"
-    exposureId <- "drug_concept_id"
-    exposurePersonId <- "person_id"
-  } else {
-    exposureStartDate <- "cohort_start_date"
-    exposureEndDate <- "cohort_end_date"
-    if (cdmVersion == "4") {
-      exposureId <- "cohort_concept_id"
-    } else {
-      exposureId <- "cohort_definition_id"
-    }
-    exposurePersonId <- "subject_id"
-  }
 
+  outcomeTable <- tolower(outcomeTable)
   if (outcomeTable == "condition_era") {
     outcomeStartDate <- "condition_era_start_date"
     outcomeId <- "condition_concept_id"
@@ -223,89 +377,114 @@ runSelfControlledCohort <- function(connectionDetails,
     outcomePersonId <- "subject_id"
   }
 
-  # Check if connection already open:
-  if ("conn" %in% names(connectionDetails)) {
-    conn <- connectionDetails$conn
-  } else {
-    conn <- DatabaseConnector::connect(connectionDetails)
-    on.exit(DatabaseConnector::disconnect(conn))
+  if (resultsTable != "#results") {
+    if (is.null(resultsDatabaseSchema))
+      stop("Results table is not temporary and resultsDatabaseSchema is not set")
+
+    resultsTable <- SqlRender::render("@results_database_schema.@results_table",
+                                      results_database_schema = resultsDatabaseSchema,
+                                      results_table = resultsTable)
   }
 
-  DatabaseConnector::insertTable(connection = conn,
-                                 tableName = "#scc_outcome_ids",
-                                 data = data.frame(outcome_id = outcomeIds),
-                                 tempTable = TRUE)
-  DatabaseConnector::insertTable(connection = conn,
-                                 tableName = "#scc_exposure_ids",
-                                 data = data.frame(exposure_id = exposureIds),
-                                 tempTable = TRUE)
+  # Check if connection already open:
+  if ("conn" %in% names(connectionDetails)) {
+    connection <- connectionDetails$conn
+  } else {
+    connection <- DatabaseConnector::connect(connectionDetails)
+    on.exit(DatabaseConnector::disconnect(connection))
+  }
+
+  if (outcomeIds != '') {
+    DatabaseConnector::insertTable(connection = connection,
+                                   tableName = "#scc_outcome_ids",
+                                   data = data.frame(outcome_id = outcomeIds),
+                                   tempTable = TRUE)
+  }
+
+  computeSccRiskWindows(connection = connection,
+                        cdmDatabaseSchema = cdmDatabaseSchema,
+                        cdmVersion = cdmVersion,
+                        oracleTempSchema = oracleTempSchema,
+                        exposureIds = exposureIds,
+                        exposureDatabaseSchema = cdmDatabaseSchema,
+                        exposureTable = exposureTable,
+                        firstExposureOnly = TRUE,
+                        minAge = minAge,
+                        maxAge = maxAge,
+                        studyStartDate = studyStartDate,
+                        studyEndDate = studyEndDate,
+                        addLengthOfExposureExposed = addLengthOfExposureExposed,
+                        riskWindowStartExposed = riskWindowStartExposed,
+                        riskWindowEndExposed = riskWindowEndExposed,
+                        addLengthOfExposureUnexposed = addLengthOfExposureUnexposed,
+                        riskWindowEndUnexposed = riskWindowEndUnexposed,
+                        riskWindowStartUnexposed = riskWindowStartUnexposed,
+                        hasFullTimeAtRisk = hasFullTimeAtRisk,
+                        washoutPeriod = washoutPeriod,
+                        followupPeriod = followupPeriod,
+                        riskWindowsTable = riskWindowsTable,
+                        resultsDatabaseSchema = resultsDatabaseSchema)
+
+  if (riskWindowsTable != "#risk_windows") {
+    riskWindowsTable <- SqlRender::render("@results_database_schema.@risk_windows_table",
+                                          results_database_schema = resultsDatabaseSchema,
+                                          risk_windows_table = riskWindowsTable)
+
+  }
 
   renderedSql <- SqlRender::loadRenderTranslateSql(sqlFilename = "Scc.sql",
                                                    packageName = "SelfControlledCohort",
-                                                   dbms = connectionDetails$dbms,
+                                                   dbms = connection@dbms,
                                                    oracleTempSchema = oracleTempSchema,
-                                                   cdm_database_schema = cdmDatabaseSchema,
-                                                   exposure_ids = exposureIds,
                                                    outcome_ids = outcomeIds,
-                                                   exposure_database_schema = exposureDatabaseSchema,
-                                                   exposure_table = exposureTable,
-                                                   exposure_start_date = exposureStartDate,
-                                                   exposure_end_date = exposureEndDate,
-                                                   exposure_id = exposureId,
-                                                   exposure_person_id = exposurePersonId,
                                                    outcome_database_schema = outcomeDatabaseSchema,
                                                    outcome_table = outcomeTable,
                                                    outcome_start_date = outcomeStartDate,
                                                    outcome_id = outcomeId,
                                                    outcome_person_id = outcomePersonId,
-                                                   first_exposure_only = firstExposureOnly,
                                                    first_outcome_only = firstOutcomeOnly,
-                                                   min_age = minAge,
-                                                   max_age = maxAge,
-                                                   study_start_date = studyStartDate,
-                                                   study_end_date = studyEndDate,
-                                                   add_length_of_exposure_exposed = addLengthOfExposureExposed,
-                                                   risk_window_start_exposed = riskWindowStartExposed,
-                                                   risk_window_end_exposed = riskWindowEndExposed,
-                                                   add_length_of_exposure_unexposed = addLengthOfExposureUnexposed,
-                                                   risk_window_end_unexposed = riskWindowEndUnexposed,
-                                                   risk_window_start_unexposed = riskWindowStartUnexposed,
-                                                   has_full_time_at_risk = hasFullTimeAtRisk,
-                                                   washout_window = washoutPeriod,
-                                                   followup_window = followupPeriod,
-                                                   compute_tar_distribution = computeTarDistribution)
+                                                   risk_windows_table = riskWindowsTable,
+                                                   results_table = resultsTable)
   ParallelLogger::logInfo("Retrieving counts from database")
-  DatabaseConnector::executeSql(conn, renderedSql)
+  DatabaseConnector::executeSql(connection, renderedSql)
+
+  if (computeTarDistribution) {
+    ParallelLogger::logInfo("Computing time at risk distribution statistics")
+    renderedSql <- SqlRender::loadRenderTranslateSql(sqlFilename = "SccRiskWindowStats.sql",
+                                                     packageName = "SelfControlledCohort",
+                                                     dbms = connection@dbms,
+                                                     oracleTempSchema = oracleTempSchema,
+                                                     outcome_ids = outcomeIds,
+                                                     outcome_database_schema = outcomeDatabaseSchema,
+                                                     outcome_table = outcomeTable,
+                                                     outcome_start_date = outcomeStartDate,
+                                                     outcome_id = outcomeId,
+                                                     outcome_person_id = outcomePersonId,
+                                                     first_outcome_only = firstOutcomeOnly,
+                                                     risk_windows_table = riskWindowsTable)
+    DatabaseConnector::executeSql(connection, renderedSql)
+    tarStats <- DatabaseConnector::renderTranslateQuerySql(connection, "SELECT * FROM #tar_stats")
+  }
+
   # Fetch results from server:
-  renderedSql <- SqlRender::loadRenderTranslateSql(sqlFilename = "LoadMergedResults.sql",
-                                                   packageName = "SelfControlledCohort",
-                                                   dbms = connectionDetails$dbms,
-                                                   oracleTempSchema = oracleTempSchema,
-                                                   compute_tar_distribution = computeTarDistribution)
-  estimates <- DatabaseConnector::querySql(conn, renderedSql)
-
-  colnames(estimates) <- SqlRender::snakeCaseToCamelCase(colnames(estimates))
-
-  # Drop temp table:
-  sql <- "TRUNCATE TABLE #results; DROP TABLE #results; {@compute_tar_distribution} ? {TRUNCATE TABLE #tar_stats; DROP TABLE #tar_stats;}"
-  DatabaseConnector::renderTranslateExecuteSql(conn,
-                                               sql,
-                                               progressBar = FALSE,
-                                               reportOverallTime = FALSE,
-                                               oracleTempSchema = oracleTempSchema,
-                                               compute_tar_distribution = computeTarDistribution)
+  ParallelLogger::logInfo("Fetching results")
+  estimates <- DatabaseConnector::renderTranslateQuerySql(connection,
+                                                          "SELECT * FROM @results_table",
+                                                          results_table = resultsTable,
+                                                          oracleTempSchema = oracleTempSchema,
+                                                          snakeCaseToCamelCase = TRUE)
 
   if (nrow(estimates) > 0) {
     ParallelLogger::logInfo("Computing incidence rate ratios and exact confidence intervals")
     zeroCountIdx <- estimates$numOutcomesExposed == 0 & estimates$numOutcomesUnexposed == 0
     if (any(zeroCountIdx)) {
-      zeroCountRows <- estimates[zeroCountIdx, ]
+      zeroCountRows <- estimates[zeroCountIdx,]
       zeroCountRows$irr <- NA
       zeroCountRows$irrLb95 <- 0
       zeroCountRows$irrUb95 <- Inf
     }
     if (any(!zeroCountIdx)) {
-      estimates <- estimates[!zeroCountIdx, ]
+      estimates <- estimates[!zeroCountIdx,]
       batches <- ceiling(nrow(estimates) / 10000)
       estimates <- split(estimates, rep_len(1:batches, nrow(estimates)))
       cluster <- ParallelLogger::makeCluster(computeThreads)
@@ -321,8 +500,20 @@ runSelfControlledCohort <- function(connectionDetails,
       }
     }
     estimates$logRr <- log(estimates$irr)
-    estimates$seLogRr <- (log(estimates$irrUb95) - log(estimates$irrLb95))/(2 * qnorm(0.975))
+    estimates$seLogRr <- (log(estimates$irrUb95) - log(estimates$irrLb95)) / (2 * qnorm(0.975))
   }
+  # Drop temp tables:
+  ParallelLogger::logInfo("Cleaning up intermedate tables")
+  sql <- SqlRender::loadRenderTranslateSql(sqlFilename = "CleanupTables.sql",
+                                           packageName = "SelfControlledCohort",
+                                           dbms = connection@dbms,
+                                           oracleTempSchema = oracleTempSchema,
+                                           outcome_ids = outcomeIds,
+                                           exposure_ids = exposureIds,
+                                           results_table = resultsTable,
+                                           compute_tar_distribution = computeTarDistribution)
+  DatabaseConnector::executeSql(connection, sql)
+
   delta <- Sys.time() - start
   ParallelLogger::logInfo(paste("Performing SCC analysis took", signif(delta, 3), attr(delta, "units")))
 
@@ -330,6 +521,11 @@ runSelfControlledCohort <- function(connectionDetails,
                  exposureIds = exposureIds,
                  outcomeIds = outcomeIds,
                  call = match.call())
+
+  if (computeTarDistribution) {
+    result$tarStats <- tarStats
+  }
+
   class(result) <- "sccResults"
   return(result)
 }